pyproject.toml
[tool.poetry]
name = "nhanes-gcp"
version = "0.1.0"
description = "ELT of NHANES data and processes for analyzing it"
authors = ["B. Ross Katz", "Alexander J. Titus"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
pandas = "^2.1.4"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
lxml = "^5.1.0"
google-cloud-storage = "^2.14.0"
google-cloud-bigquery = "^3.15.0"
pandas-gbq = "^0.20.0"
google-api-python-client = "^2.113.0"
duckdb = "^0.9.2"
matplotlib = "^3.8.2"
duckdb-engine = "^0.10"
python-dotenv = "^1.0.0"
dbt = "^1.0.0.35.4"
dbt-core = "^1.7.4"
dbt-bigquery = "^1.7.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
black = "^23.12.1"
ruff = "^0.1.13"
isort = "^5.13.2"
pytest = "^7.4.4"
ipykernel = "^6.26.0"
sqlfluff = "^2.3.5"
jupyterlab-code-formatter = "^2.2.1"
notebook = "^7.0.6"
jupysql = "^0.10.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
